+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Force all persons to be allocated to all entities
   - Improve error messages
   - Detect unexpected entity errors first
+- Document new simulation and entities constructors.
 
 ## 14.1.3 - [#541](https://github.com/openfisca/openfisca-core/pull/541)